Remove the Ctrl+C handling thread
It was causing all the inputs to be responded to inconsistently, which was not worth it.

The app is way more responsive now.

I assume that running event::poll() from two different places caused ~half the events to be read by the "wrong" event handler, so they just got dropped.